+++ lot.php
@@ -1,6 +1,7 @@
 <?php
 require ('functions.php');
 require ('all_lots.php');
+require ('all_bets.php');
 
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
@@ -11,33 +12,6 @@ if (isset($_GET['id'])) {
         http_response_code(404);
         header("Location: /pages/404.html");
     };
-};
-
-// ставки пользователей, которыми надо заполнить таблицу
-$bets = [
-    ['name' => 'Иван', 'price' => 11500, 'ts' => strtotime('-' . rand(1, 50) .' minute')],
-    ['name' => 'Константин', 'price' => 11000, 'ts' => strtotime('-' . rand(1, 18) .' hour')],
-    ['name' => 'Евгений', 'price' => 10500, 'ts' => strtotime('-' . rand(25, 50) .' hour')],
-    ['name' => 'Семён', 'price' => 10000, 'ts' => strtotime('last week')]
-];
-
-function time_bet($ts) {
-    $now = strtotime('now');
-    $difference_hours = ($now - $ts)/3600;
-    if ($difference_hours > 24) {
-        $time = date('d.m.y', $ts) . ' в ' . date('H.i', $ts);
-    }
-    else {
-        if ($difference_hours < 1) {
-            $time = date('i', ($now - $ts)) . ' минут назад';
-        }
-        else {
-            $time_zone = date('Z'); 
-            $time = date('G', ($now - $ts - $time_zone)) . ' часов назад';
-        }
-    }
-    
-    return $time;
 };
 
 $data_page = get_template ('lot', ['bets' => $bets, 'lot_item' => $lot_item ]);
